wip: [stix2 export] Exporting confidence level tags in the STIX 2.1 confidence field
